修正ad已被删除不能再删除

+++ oomall/ad/src/main/java/xmu/yida/ad/dao/AdDao.java
@@ -41,6 +41,13 @@ public class AdDao {
     }
 
     public boolean deleteAdById(Integer id){
+        Ad ad=adMapper.findAdById(id);
+        if(ad==null){
+            return false;
+        }
+        if(ad.getBeDeleted()){
+            return false;
+        }
         return adMapper.deleteAdById(id);
     }
